LayoutTests/platform/glib/http/tests/websocket/tests/hybi/handshake-fail-by-non-ascii-header-value-sec-websocket-extensions-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-non-ascii-header-value-sec-websocket-extensions' failed: Server requested unsupported extension
Test whether WebSocket handshake fails if the server sends a WebSocket extensions header with non-ascii characters in its value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
